Refactor: Size 타입 AddButton 에서서 Chips로 이동. ColorDotButton 이름수정
