+++ src/components/ui/AddButton.tsx
@@ -1,11 +1,6 @@
 import Image from 'next/image';
 
-export enum Size {
-	Xsmall = 6,
-	Small = 8,
-	Medium = 14.5,
-	Large = 16,
-}
+import { Size } from './Chips/Chips';
 
 export default function AddButton({ size }: { size: Size }): React.JSX.Element {
 	return (
